Add a few words on HasSubtitles
OSDb's method for this is actually named TryUploadSubtitles, although it
does not require any form of file upload. It *is* a little weird.

I haven't looked into it too much yet. However, I am thinking that the
XML field names we send here *are* case-sensitive because the API always
replies with a "false", when checking wheteher a freshly found subtitle
already exists...

Which would mean that queries parameter names are all lowercased (per
the docs), but that, for server replies, the same fields are camel-case.
Well done! :sob: