Cleaning Subtitle stype
Subtitle.toUploadParams does less, and does not include the client's
Token automatically – which allows us to remove the only parameter to
this method.

TryUploadSubtitle now ignores the <data> returned by OSDB: because we
need to be able to map a concrete type to the value, but the server
insists on returning either <array> of structs, or plain <struct>s under
the same name.

One way to solve this second problem would be a two-pass decoding... :fearful: