Use HTTPS and cleanup API style
OSDB was initially not available over HTTPS, which was bad but they
fixed that recently. Now you can grab subtitles and leak a little less
information on what you're watching.

The "cleanup" of the API consists of the following:

 - Rename some methods related to IMDB,
 - Use go style attributes ("ID" instead of "id"),
 - Use go style variable case ("fooBar" instead of "foo_bar").

This is a breaking change, especially if you used IMDB methods but it
does not affect you as long as you use to the 0.2 release.